LayoutTests/imported/w3c/web-platform-tests/css/cssom/page-descriptors-expected.txt


FAIL Page descriptors shouldn't be exposed to CSS Style declarations assert_equals: computed style should not have size property expected (undefined) undefined but got (string) ""
